server/requirements.txt
fastapi
uvicorn[standard]
pyngrok
python-dotenv